Updating pom.xml to v1.5.0
A Major rewrite of the codebase is coming, and we'll be updating to v1.5.0 once that is done.

This rewrite will focus on evaluating the validity of the code from an OOP-perspective. This means that Utility-classes wil be rewritten to State-based classes.

These changes will also affect loading the data within the config.yml.

Expect the update to take about a week, give or take a day or three.