Checking if game is eligible to start
Previously it was possible to start a game with only one player in it, and no players assigned to either team. Now at least one player needs to be assigned to both teams, or at least two players need to have joined the game and the playerroles must be random.