Unload and Destroy Worlds Asynchronously
Rather than waiting for the server to stop, we now delete unload worlds immediately after the /stopgame command is given. Except if the universe name is "world".

Furthermore, we also Asynchronously delete the associated world files, if, and only if, the universe is marked for destruction. This way players can preserve their worlds if they so choose to, and can immediately recreate a game with the same name, if they choose to delete their world.